pyproject.toml
[project]
name = "ai-medical-chatbot-with-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.23",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
